+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Feb 20, 2024]
+- add: view for batches on hold, including count per presentation and total
+
+
 ## [Feb 17, 2024]
 - add: ability to edit batch using a modal view
 - add: ability to edit target dates using a modal view